Revert "PR: ItemModeScoreBoard기능 추가"
