+++ src/main/java/game/container/BlockGenerator.java
@@ -73,7 +73,7 @@ public class BlockGenerator {
 	public BlockController createBlock() {
 		BlockController curBlock = blockQueue.poll();
 		initNewBlockPos(curBlock, 0, 5);
-		
+		GameManager_NormalMode.getInstance().curBlock = curBlock;
 		return curBlock;
 	}
