+++ src/main/java/save/KeySetting.json
@@ -1,1 +1,1 @@
-{"stop":84,"left":37,"downBlock":40,"right":39,"turnBlock":38,"oneTimeDown":32,"escape":27}
+{"stop":84,"left":37,"downBlock":40,"right":39,"turnBlock":38,"oneTimeDown":32}
