--- src/main/java/save/NoItemScoreList.json
@@ -1,1 +1,0 @@
-{"mode4":"easy","mode3":"easy","score2":20,"score3":20,"score4":20,"name4":"40","name3":"a","score0":20,"score1":20,"name2":"40","mode2":"easy","name1":"12","mode1":"easy","name0":"13","mode0":"easy"}
