+++ src/main/java/save/ScoreList.json
@@ -1,0 +1,1 @@
+{}
