+++ src/main/java/scoreBoard/User.java
@@ -6,14 +6,9 @@ public class User implements Comparable<User>{
 
     private int score;
 
-    private String mode;
-
-
-
-    public User(String name, int score,String mode) {
+    public User(String name, int score) {
         this.name = name;
         this.score = score;
-        this.mode = mode;
     }
 
     public String getName() {
@@ -30,13 +25,6 @@ public class User implements Comparable<User>{
 
     public void setScore(int score) {
         this.score = score;
-    }
-    public String getMode() {
-        return mode;
-    }
-
-    public void setMode(String mode) {
-        this.mode = mode;
     }
